Fixed the issues from previous commit
