+++ include/lions/firewall/routing.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
+#include <lions/firewall/array_functions.h>
 #include <lions/firewall/common.h>
 #include <lions/firewall/queue.h>
 
@@ -247,7 +249,7 @@ static fw_routing_err_t fw_routing_find_route(fw_routing_table_t *table,
                                       uint8_t num_calls)
 {
     fw_routing_entry_t *match = NULL;
-    for (uint16_t i = 0; i < table->size; ++i) {
+    for (uint16_t i = 0; i < table->size; i++) {
         fw_routing_entry_t *entry = table->entries + i;
 
         /* ip is part of subnet */
@@ -317,16 +319,12 @@ static fw_routing_err_t fw_routing_table_add_route(fw_routing_table_t *table,
         }
     }
 
-    if (table->size == table->capacity) {
-        return ROUTING_ERR_FULL;
-    }
-
     fw_routing_entry_t *empty_slot = table->entries + table->size;
-    table->size++;
     empty_slot->interface = interface;
     empty_slot->ip = subnet_mask(subnet) & ip;
     empty_slot->subnet = subnet;
     empty_slot->next_hop = next_hop;
+    table->size++;
 
     return ROUTING_ERR_OKAY;
 }
